refactor: Remove option to skip owner setup
