+++ cypress/e2e/12-canvas-actions.cy.ts
@@ -6,16 +6,25 @@ import {
 	SET_NODE_NAME,
 	IF_NODE_NAME,
 	HTTP_REQUEST_NODE_NAME,
+	DEFAULT_USER_EMAIL,
+	DEFAULT_USER_PASSWORD,
 } from './../constants';
 import { WorkflowPage as WorkflowPageClass } from '../pages/workflow';
+import { randFirstName, randLastName } from '@ngneat/falso';
+
+const email = DEFAULT_USER_EMAIL;
+const password = DEFAULT_USER_PASSWORD;
+const firstName = randFirstName();
+const lastName = randLastName();
 
 const WorkflowPage = new WorkflowPageClass();
 describe('Canvas Actions', () => {
 	before(() => {
-		cy.skipSetup();
+		cy.setup({ email, firstName, lastName, password });
 	});
 
 	beforeEach(() => {
+		cy.signin({ email, password });
 		WorkflowPage.actions.visit();
 	});
