+++ cypress/e2e/26-resource-locator.cy.ts
@@ -1,3 +1,5 @@
+import { randFirstName, randLastName } from '@ngneat/falso';
+import { DEFAULT_USER_EMAIL, DEFAULT_USER_PASSWORD } from '../constants';
 import { WorkflowPage, NDV, CredentialsModal } from '../pages';
 
 const workflowPage = new WorkflowPage();
@@ -7,12 +9,18 @@ const credentialsModal = new CredentialsModal();
 const NO_CREDENTIALS_MESSAGE = 'Please add your credential';
 const INVALID_CREDENTIALS_MESSAGE = 'Please check your credential';
 
+const email = DEFAULT_USER_EMAIL;
+const password = DEFAULT_USER_PASSWORD;
+const firstName = randFirstName();
+const lastName = randLastName();
+
 describe('Resource Locator', () => {
 	before(() => {
-		cy.skipSetup();
+		cy.setup({ email, firstName, lastName, password });
 	});
 
 	beforeEach(() => {
+		cy.signin({ email, password });
 		workflowPage.actions.visit();
 	});
