+++ cypress/e2e/5-ndv.cy.ts
@@ -1,15 +1,23 @@
+import { randFirstName, randLastName } from '@ngneat/falso';
+import { DEFAULT_USER_EMAIL, DEFAULT_USER_PASSWORD } from '../constants';
 import { WorkflowPage, NDV } from '../pages';
 import { v4 as uuid } from 'uuid';
 
 const workflowPage = new WorkflowPage();
 const ndv = new NDV();
 
+const email = DEFAULT_USER_EMAIL;
+const password = DEFAULT_USER_PASSWORD;
+const firstName = randFirstName();
+const lastName = randLastName();
+
 describe('NDV', () => {
 	before(() => {
-		cy.skipSetup();
+		cy.setup({ email, firstName, lastName, password });
 	});
 
 	beforeEach(() => {
+		cy.signin({ email, password });
 		workflowPage.actions.visit();
 		workflowPage.actions.renameWorkflow(uuid());
 		workflowPage.actions.saveWorkflowOnButtonClick();
@@ -210,7 +218,7 @@ describe('NDV', () => {
 		ndv.getters.inputRunSelector()
 			.find('input')
 			.should('include.value', '1 of 2 (6 items)');
-		
+
 		// unlink again
 		ndv.actions.toggleInputRunLinking();
 		ndv.getters.inputTbodyCell(1, 0).click(); // remove tooltip
