+++ cypress/e2e/6-code-node.cy.ts
@@ -1,12 +1,24 @@
 import { WorkflowPage as WorkflowPageClass } from '../pages/workflow';
 import { NDV } from '../pages/ndv';
+import { randFirstName, randLastName } from '@ngneat/falso';
+import { DEFAULT_USER_EMAIL, DEFAULT_USER_PASSWORD } from '../constants';
 
 const WorkflowPage = new WorkflowPageClass();
 const ndv = new NDV();
 
+const email = DEFAULT_USER_EMAIL;
+const password = DEFAULT_USER_PASSWORD;
+const firstName = randFirstName();
+const lastName = randLastName();
+
 describe('Code node', () => {
 	before(() => {
-		cy.skipSetup();
+		cy.setup({ email, firstName, lastName, password });
+	});
+
+	beforeEach(() => {
+		cy.signin({ email, password });
+		WorkflowPage.actions.visit();
 	});
 
 	it('should execute the placeholder in all-items mode successfully', () => {
@@ -20,7 +32,6 @@ describe('Code node', () => {
 	});
 
 	it('should execute the placeholder in each-item mode successfully', () => {
-		WorkflowPage.actions.visit();
 		WorkflowPage.actions.addInitialNodeToCanvas('Manual');
 		WorkflowPage.actions.addNodeToCanvas('Code');
 		WorkflowPage.actions.openNode('Code');
