+++ packages/cli/src/config/types.d.ts
@@ -80,7 +80,6 @@ type ExceptionPaths = {
 	'nodes.exclude': string[] | undefined;
 	'nodes.include': string[] | undefined;
 	'userManagement.isInstanceOwnerSetUp': boolean;
-	'userManagement.skipInstanceOwnerSetup': boolean;
 };
 
 // -----------------------------------
