+++ packages/editor-ui/src/plugins/i18n/locales/en.json
@@ -95,15 +95,11 @@
 	"auth.setup.createAccount": "Create account",
 	"auth.setup.goBack": "Go back",
 	"auth.setup.next": "Next",
-	"auth.setup.ownerAccountBenefits": "Setting up an owner account allows you to invite others, and prevents people using n8n without an account",
 	"auth.setup.settingUpOwnerError": "Problem setting up owner",
 	"auth.setup.setupConfirmation.concatEntities": "{workflows} and {credentials}",
 	"auth.setup.setupConfirmation.credentials": "{count} credential | {count} credentials",
 	"auth.setup.setupConfirmation.existingWorkflows": "{count} existing workflow | {count} existing workflows",
 	"auth.setup.setupOwner": "Set up owner account",
-	"auth.setup.skipOwnerSetupQuestion": "Skip owner account setup?",
-	"auth.setup.skipSetup": "Skip setup",
-	"auth.setup.skipSetupTemporarily": "Skip setup for now",
 	"auth.signin": "Sign in",
 	"auth.signin.error": "Problem logging in",
 	"auth.signout": "Sign out",
