+++ packages/editor-ui/src/router.ts
@@ -530,14 +530,7 @@ export const routes = [
 					},
 					permissions: {
 						allow: {
-							role: [ROLE.Default, ROLE.Owner],
-						},
-						deny: {
-							shouldDeny: () => {
-								const settingsStore = useSettingsStore();
-
-								return !(settingsStore.isCloudDeployment || settingsStore.isDesktopDeployment);
-							},
+							role: [ROLE.Owner],
 						},
 					},
 				},
@@ -635,7 +628,7 @@ export const routes = [
 					},
 					permissions: {
 						allow: {
-							role: [ROLE.Default, ROLE.Owner],
+							role: [ROLE.Owner],
 						},
 						deny: {
 							role: [ROLE.Member],
@@ -655,7 +648,7 @@ export const routes = [
 					},
 					permissions: {
 						allow: {
-							role: [ROLE.Default, ROLE.Owner],
+							role: [ROLE.Owner],
 						},
 						deny: {
 							shouldDeny: () => {
@@ -697,7 +690,7 @@ export const routes = [
 				meta: {
 					permissions: {
 						allow: {
-							role: [ROLE.Default, ROLE.Owner],
+							role: [ROLE.Owner],
 						},
 						deny: {
 							role: [ROLE.Member],
